Added --find-missing flag to check for missing data in buckets.
If this flag is set (default: off), it will not execute the recipe. Instead, it will iterate through all of the FilePatterns to see if any data is missing. If a path does not exist, it will print it out.

This provides the user a useful sanity check. This way, they can find out early if the pipeline will break due to missing files.

Change-Id: Ie5cf0e5cd6206c1529e877585576638cc6fb611a