Reduced Chunking size for model-levels-to-zarr. (#31)
In response to Issue
https://github.com/google-research/arco-era5/issues/28, where it was
identified that the chunk size in the model-level Zarr data was
excessively large, I have made significant improvements. The existing
chunk size was in the order of gigabytes. To address this, I introduced
a modification in the `target_chunks` method of the `XarrayZarrRecipe`.

By incorporating the `hybrid: 1` parameter in the `target_chunks`
method, I was able to split each chunk per level.