`level` change in the example of weather-sp. (#41)
In the
[Readme.md](https://github.com/google-research/arco-era5/blob/main/raw/README.md)
file of the raw directory, the example for `weather-sp` includes an
attribute called `level`. However, upon examining the existing data (one
of the file:
gs://gcp-public-data-arco-era5/raw/ERA5GRIB/HRES/Month/2021/202112_hres_soil.grb2),
I found that the correct attribute is **`typeOfLevel`**. This PR updates
the attribute name to `typeOfLevel` in the example to reflect the
accurate data structure.