Remove inclusive from the source_data. (#80)
Remove `inclusive=left` from the `daily_date_iterator` function of the
`source_data.py` file. This will fixed the issue #78.

Reason: while running the script on the monthly basis this didn't
include the last date of the month due to the `inclusive=left` so after
removing this it will added the last day's data, too.